+++ R/names.R
@@ -26,7 +26,7 @@ check_names <- function(x, names = character(0), exclusive = FALSE, order = FALS
                         unique = FALSE, complete = TRUE,
                         x_name = substitute(x),
                         error = TRUE) {
-  x_name <- deparse_x_name(x_name)
+  x_name <- chk_deparse(x_name)
   
   check_flag_internal(unique)
   check_vector(names, "", unique = unique)
@@ -47,13 +47,13 @@ check_names <- function(x, names = character(0), exclusive = FALSE, order = FALS
   }
   
   if (complete && length(setdiff(names, x_names)))
-    chk_fail(x_name, " names must include ", cc(setdiff(names, x_names), "and") , error = error)
+    chk_fail(x_name, " names must include ", cc_and(setdiff(names, x_names)) , error = error)
   
   if (exclusive && length(setdiff(x_names, names)))
-    chk_fail(x_name, " names must not include ", cc(setdiff(x_names, names), "or") , error = error)
+    chk_fail(x_name, " names must not include ", cc_or(setdiff(x_names, names)) , error = error)
   
   if(order && !identical(intersect(names, x_names), intersect(x_names, names)))
-    chk_fail(x_name, " names must include ", cc(names, "and"), " in that order", error = error)
+    chk_fail(x_name, " names must include ", cc_and(names), " in that order", error = error)
   
   invisible(x)
 }